+++ shuup/admin/modules/categories/__init__.py
@@ -86,5 +86,5 @@ class CategoryModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(Category) | get_default_model_permissions(File)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(Category, "shuup_admin:category", object, kind)