+++ shuup/admin/modules/contact_groups/__init__.py
@@ -52,5 +52,5 @@ class ContactGroupModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(ContactGroup)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(ContactGroup, "shuup_admin:contact_group", object, kind)