+++ shuup/admin/modules/manufacturers/__init__.py
@@ -43,5 +43,5 @@ class ManufacturerModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(Manufacturer)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(Manufacturer, "shuup_admin:manufacturer", object, kind)