+++ shuup/admin/modules/service_providers/__init__.py
@@ -56,5 +56,5 @@ class ServiceProviderModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(ServiceProvider)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(ServiceProvider, "shuup_admin:service_provider", object, kind)