+++ shuup/admin/modules/services/__init__.py
@@ -60,7 +60,7 @@ class ServiceModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(PaymentMethod) | get_default_model_permissions(ShippingMethod)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(self.model, self.url_name_prefix, object, kind)