+++ shuup/admin/modules/shops/__init__.py
@@ -68,5 +68,5 @@ class ShopModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(Shop) | get_default_model_permissions(File)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(Shop, "shuup_admin:shop", object, kind)