+++ shuup/admin/modules/suppliers/__init__.py
@@ -43,5 +43,5 @@ class SupplierModule(AdminModule):
     def get_required_permissions(self):
         return get_default_model_permissions(Supplier)
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(Supplier, "shuup_admin:supplier", object, kind)