+++ shuup/front/apps/carousel/admin_module/__init__.py
@@ -56,5 +56,5 @@ class CarouselModule(AdminModule):
             get_default_model_permissions(Product)
         )
 
-    def get_model_url(self, object, kind):
+    def get_model_url(self, object, kind, shop=None):
         return derive_model_url(Carousel, "shuup_admin:carousel", object, kind)